+++ packages/core/src/LunarServiceProvider.php
@@ -20,13 +20,10 @@ use Lunar\Core\Addons\Manifest;
 use Lunar\Core\Auth\Manifest as AccessControlManifest;
 use Lunar\Core\Base\AttributeManifest;
 use Lunar\Core\Base\AttributeManifestInterface;
-use Lunar\Core\Base\CartLineModifiers;
-use Lunar\Core\Base\CartModifiers;
 use Lunar\Core\Base\CartSessionInterface;
 use Lunar\Core\Base\DiscountManagerInterface;
 use Lunar\Core\Base\ModelManifest;
 use Lunar\Core\Base\ModelManifestInterface;
-use Lunar\Core\Base\OrderModifiers;
 use Lunar\Core\Base\OrderReferenceGenerator;
 use Lunar\Core\Base\OrderReferenceGeneratorInterface;
 use Lunar\Core\Base\PaymentManagerInterface;
@@ -34,7 +31,6 @@ use Lunar\Core\Base\PricingManagerInterface;
 use Lunar\Core\Base\ProvidesTelemetryInsights;
 use Lunar\Core\Base\ShippingManifest;
 use Lunar\Core\Base\ShippingManifestInterface;
-use Lunar\Core\Base\ShippingModifiers;
 use Lunar\Core\Base\StorefrontSessionInterface;
 use Lunar\Core\Base\TaxManagerInterface;
 use Lunar\Core\Base\TelemetryInsights;
@@ -78,6 +74,10 @@ use Lunar\Core\Models\ProductVariant;
 use Lunar\Core\Models\Staff;
 use Lunar\Core\Models\Transaction;
 use Lunar\Core\Models\Url;
+use Lunar\Core\Modifiers\CartLineModifiers;
+use Lunar\Core\Modifiers\CartModifiers;
+use Lunar\Core\Modifiers\OrderModifiers;
+use Lunar\Core\Modifiers\ShippingModifiers;
 use Lunar\Core\Observers\AddressObserver;
 use Lunar\Core\Observers\CartLineObserver;
 use Lunar\Core\Observers\ChannelObserver;
