+++ packages/core/src/Modifiers/CartLineModifiers.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Lunar\Core\Base;
+namespace Lunar\Core\Modifiers;
 
 use Illuminate\Support\Collection;
 
-class CartModifiers
+class CartLineModifiers
 {
     protected Collection $modifiers;
