+++ packages/core/src/Modifiers/CartModifiers.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Lunar\Core\Base;
+namespace Lunar\Core\Modifiers;
 
 use Illuminate\Support\Collection;
 
-class CartLineModifiers
+class CartModifiers
 {
     protected Collection $modifiers;
