+++ packages/table-rate-shipping/src/ShippingServiceProvider.php
@@ -4,12 +4,12 @@ namespace Lunar\Shipping;
 
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Support\ServiceProvider;
-use Lunar\Core\Base\ShippingModifiers;
 use Lunar\Core\Facades\Discounts;
 use Lunar\Core\Facades\ModelManifest;
 use Lunar\Core\Models\CustomerGroup;
 use Lunar\Core\Models\Order;
 use Lunar\Core\Models\Product;
+use Lunar\Core\Modifiers\ShippingModifiers;
 use Lunar\Shipping\DiscountTypes\ShippingDiscount;
 use Lunar\Shipping\Interfaces\ShippingMethodManagerInterface;
 use Lunar\Shipping\Managers\PostcodeManager;
