+++ packages/upgrade/src/Rector/LunarSetList.php
@@ -47,10 +47,10 @@ final class LunarSetList
         'Lunar\\Base\\AttributeManifest' => 'Lunar\\Core\\Base\\AttributeManifest',
         'Lunar\\Base\\AttributeManifestInterface' => 'Lunar\\Core\\Base\\AttributeManifestInterface',
         'Lunar\\Base\\BaseModel' => 'Lunar\\Core\\Base\\BaseModel',
-        'Lunar\\Base\\CartLineModifier' => 'Lunar\\Core\\Base\\CartLineModifier',
-        'Lunar\\Base\\CartLineModifiers' => 'Lunar\\Core\\Base\\CartLineModifiers',
-        'Lunar\\Base\\CartModifier' => 'Lunar\\Core\\Base\\CartModifier',
-        'Lunar\\Base\\CartModifiers' => 'Lunar\\Core\\Base\\CartModifiers',
+        'Lunar\\Base\\CartLineModifier' => 'Lunar\\Core\\Modifiers\\CartLineModifier',
+        'Lunar\\Base\\CartLineModifiers' => 'Lunar\\Core\\Modifiers\\CartLineModifiers',
+        'Lunar\\Base\\CartModifier' => 'Lunar\\Core\\Modifiers\\CartModifier',
+        'Lunar\\Base\\CartModifiers' => 'Lunar\\Core\\Modifiers\\CartModifiers',
         'Lunar\\Base\\CartSessionInterface' => 'Lunar\\Core\\Base\\CartSessionInterface',
         'Lunar\\Base\\Casts\\AsAttributeData' => 'Lunar\\Core\\Casts\\AsAttributeData',
         'Lunar\\Base\\Casts\\CouponString' => 'Lunar\\Core\\Casts\\CouponString',
@@ -78,8 +78,8 @@ final class LunarSetList
         'Lunar\\Base\\Migration' => 'Lunar\\Core\\Base\\Migration',
         'Lunar\\Base\\ModelManifest' => 'Lunar\\Core\\Base\\ModelManifest',
         'Lunar\\Base\\ModelManifestInterface' => 'Lunar\\Core\\Base\\ModelManifestInterface',
-        'Lunar\\Base\\OrderModifier' => 'Lunar\\Core\\Base\\OrderModifier',
-        'Lunar\\Base\\OrderModifiers' => 'Lunar\\Core\\Base\\OrderModifiers',
+        'Lunar\\Base\\OrderModifier' => 'Lunar\\Core\\Modifiers\\OrderModifier',
+        'Lunar\\Base\\OrderModifiers' => 'Lunar\\Core\\Modifiers\\OrderModifiers',
         'Lunar\\Base\\OrderReferenceGenerator' => 'Lunar\\Core\\Base\\OrderReferenceGenerator',
         'Lunar\\Base\\OrderReferenceGeneratorInterface' => 'Lunar\\Core\\Base\\OrderReferenceGeneratorInterface',
         'Lunar\\Base\\PaymentManagerInterface' => 'Lunar\\Core\\Base\\PaymentManagerInterface',
@@ -89,8 +89,8 @@ final class LunarSetList
         'Lunar\\Base\\Purchasable' => 'Lunar\\Core\\Base\\Purchasable',
         'Lunar\\Base\\ShippingManifest' => 'Lunar\\Core\\Base\\ShippingManifest',
         'Lunar\\Base\\ShippingManifestInterface' => 'Lunar\\Core\\Base\\ShippingManifestInterface',
-        'Lunar\\Base\\ShippingModifier' => 'Lunar\\Core\\Base\\ShippingModifier',
-        'Lunar\\Base\\ShippingModifiers' => 'Lunar\\Core\\Base\\ShippingModifiers',
+        'Lunar\\Base\\ShippingModifier' => 'Lunar\\Core\\Modifiers\\ShippingModifier',
+        'Lunar\\Base\\ShippingModifiers' => 'Lunar\\Core\\Modifiers\\ShippingModifiers',
         'Lunar\\Base\\StandardMediaDefinitions' => 'Lunar\\Core\\Base\\StandardMediaDefinitions',
         'Lunar\\Base\\StorefrontSessionInterface' => 'Lunar\\Core\\Base\\StorefrontSessionInterface',
         'Lunar\\Base\\TaxDriver' => 'Lunar\\Core\\Base\\TaxDriver',
